feat: signin auth with firebase api
The Auth.tsx was changed to incorporate the signin auth methods. The signin calls the firebase with the user email and passwords and set the auth:true in the context app. This lets the user opens the app.

Besides that, it is possible to signout using the firebase signout api function

A button with the navigate to the userRegister page was created to simply simulate the creating of a new user in the future