fix/refactor: userRegister registration handling
The method of persisting user data in the database has been changed. Now, it utilizes the setDoc function to record a document given its ID. This ID is the uid field from Firebase Authentication, received when creating a new user.

To retrieve data during authentication (Login), a database query is now performed to fetch the user based on the Firebase Authentication UID.

There are still issues in the code that need to be resolved, and for that, issues should be created.